requirements.txt
# FastAPI & server
fastapi==0.115.0
uvicorn==0.30.1

# AI / Face Recognition
insightface==0.7.3
onnxruntime==1.17.1
onnx==1.15.0

# Machine Learning
scikit-learn==1.5.2
numpy==1.26.4
scipy==1.13.1

# Image Processing
opencv-python-headless==4.10.0.84
Pillow==10.3.0

# Utilities
pydantic==2.8.2
python-multipart==0.0.9

# Optional (jika kamu mau logging performa GPU)
tensorflow==2.17.0